Review fixes for all the sections done
