+++ src/styles/abstracts/_functions.scss
@@ -1,5 +1,4 @@
 // Function to convert px to rem
-
 @function pxToRem($size, $base: 10) {
     @return calc($size/$base) * 1rem;
 }
