+++ src/styles/base/_typography.scss
@@ -6,10 +6,11 @@
 // Hero Section Heading
 // ==============================
 
-.hero-heading {
+.hero__heading {
     // Base font size using custom pxToRem function
     font-size: var.$hero-heading-size;
     font-weight: bold;
+    @include mix.line-clamp(4);
     & .heading--highlight {
         background-color: transparent;
         color: var.$highlight-color;
@@ -34,7 +35,8 @@
     // Base font size and weight
     font-size: func.pxToRem(32);
     font-weight: bold;
-
+    line-height: 1.2;
+    @include mix.line-clamp(3);
     // Responsive scaling for medium screens
     @include mix.respond('md') {
         font-size: func.pxToRem(40);
@@ -66,7 +68,7 @@
     // Set font family and default muted color
     font-family: var.$font-paragraph;
     color: var.$muted-text-color;
-
+    @include mix.line-clamp(5);
     // Increase font size on large screens
     @include mix.respond('sm') {
         font-size: func.pxToRem(18);
@@ -86,11 +88,19 @@
     &--travel {
         font-size: func.pxToRem(18);
         color: var.$black;
+        @include mix.line-clamp();
     }
 
     // Font adjustments for services section paragraph
     &--services {
         font-size: func.pxToRem(18);
+        @include mix.line-clamp($lines: 2);
+        @include mix.respond('lg') {
+            @include mix.line-clamp($lines: 4);
+        }
+        @include mix.respond('xl') {
+            @include mix.line-clamp($lines: 3);
+        }
     }
 
     // Font adjustments for testimonial section paragraph
@@ -98,6 +108,7 @@
         font-size: func.pxToRem(18);
         font-family: var.$font-family-base;
         color: var.$black;
+        @include mix.line-clamp($lines: 3);
         @include mix.respond('lg') {
             font-size: func.pxToRem(23);
         }
@@ -122,7 +133,8 @@
     font-size: func.pxToRem(16);
     color: var.$highlight-color;
     letter-spacing: func.pxToRem(4);
-
+    text-transform: uppercase;
+    @include mix.line-clamp();
     // Responsive scaling for larger screens
     @include mix.respond('lg') {
         font-size: func.pxToRem(24);
@@ -137,6 +149,7 @@
 .card-subheading {
     font-size: func.pxToRem(18);
     color: var.$orange;
+    @include mix.line-clamp(3);
 
     // Font adjustments for services section
     &--services {
@@ -176,32 +189,22 @@
 }
 
 // ==============================
-// Newsletter Form Input
-// ==============================
-
-.newsletter__form-input {
-    font-size: func.pxToRem(14);
-    font-weight: bold;
-    color: var.$muted-text-color;
-    @include mix.respond('lg') {
-        font-size: func.pxToRem(23);
-    }
-}
-
-// ==============================
 // Footer Section Accordion
 // ==============================
 
 .footer {
     &__accordion {
-        &-toggle {
-            font-size: func.pxToRem(23);
-            font-weight: bold;
-        }
         &-list {
             font-family: var.$font-paragraph;
             font-size: func.pxToRem(18);
             color: var.$muted-text-color;
         }
+    }
+}
+
+.js-footer__accordion {
+    &-toggle {
+        font-size: func.pxToRem(23);
+        font-weight: bold;
     }
 }
