+++ src/styles/main.scss
@@ -5,6 +5,7 @@
 @use 'base/reset';
 @use 'base/fonts';
 @use 'components/button';
+@use 'components/form';
 @use 'base/typography';
 @use 'layout/common';
 @use 'layout/header';
