+++ src/styles/vendors/_icons.scss
@@ -59,7 +59,7 @@
 // =============================================
 // Icon: Hamburger Menu (mobile nav)
 // =============================================
-.nav__hamburger-icon:before {
+.header__hamburger-icon:before {
     content: '\e906';
     color: var.$primary-color;
 }
